a separate button for the cropping file ability
