+++ src/components/MediaControlPanel/MediaControlPanel.jsx
@@ -10,17 +10,12 @@ import AudioTimeSlider from '../AudioTimeSlider/AudioTimeSlider.jsx';
 import ZoomSlider from '../ZoomSlider/ZoomSlider.jsx';
 import AudioTimeControlButtons from '../AudioTimeControlButtons/AudioTimeControlButtons.jsx';
 import VolumeSlider from '../VolumeSlider/VolumeSlider.jsx';
-import Modal from '../Modal/Modal.jsx';
-
-const MODAL_HEADER_MESSAGE = 'Please wait';
-const MODAL_BODY_MESSAGE = "We're editing your file...";
 
 export default function MediaControlPanel() {
   const dispatch = useDispatch();
 
-  const { wavesurfer, regionsPlugin, audioFile } = useContext(WaveSurferContext);
+  const { wavesurfer } = useContext(WaveSurferContext);
   const [currentTime, setCurrentTime] = useState(0);
-  const [openModal, setOpenModal] = useState(false);
 
   useEffect(() => {
     const unsubscribeTimeupdateEvent = wavesurfer.current.on('timeupdate', (time) => {
@@ -36,44 +31,6 @@ export default function MediaControlPanel() {
     }
   }, []);
 
-  const download = (blob) => {
-    const fileURL = URL.createObjectURL(blob);
-
-    const a = document.createElement('a');
-    a.href = fileURL;
-    a.download = 'cropped-audio.mp3';
-    document.body.appendChild(a);
-
-    a.click();
-    a.remove();
-    setOpenModal(false);
-  };
-
-  const cutAudio = async () => {
-    try {
-      const regions = regionsPlugin.current.getRegions();
-      if (!regions.length) return;
-
-      setOpenModal(true);
-      const { start: startTime, end: endTime } = regions[regions.length - 1];
-      const body = new FormData();
-      body.append('file', audioFile);
-      body.append('start_time', startTime);
-      body.append('end_time', endTime);
-
-      const response = await fetch('http://localhost:8000/audio/crop-audio', {
-        method: 'POST',
-        body
-      });
-
-      const blob = await response.blob();
-      download(blob);
-    } catch (err) {
-      console.error('Failure on cropping audio');
-      console.error(err);
-    }
-  };
-
   return (
     <>
       <div className={styles.controls}>
@@ -83,17 +40,11 @@ export default function MediaControlPanel() {
             <ZoomSlider />
           </div>
           <AudioTimeControlButtons />
-          <button onClick={cutAudio} className={styles.downloadBtn}>Download</button>
           <div className={styles.controlGroup}>
             <VolumeSlider />
           </div>
         </div>
       </div>
-      <Modal
-        isOpened={openModal}
-        headerMessage={MODAL_HEADER_MESSAGE}
-        bodyMessage={MODAL_BODY_MESSAGE}
-      />
     </>
   );
 }
