+++ src/components/MediaControlPanel/MediaControlPanel.module.css
@@ -21,14 +21,3 @@ input[type='range']::-webkit-slider-thumb {
   align-items: center;
   gap: 1em;
 }
-
-.download-btn {
-  background-color: #00b846;
-  border: 0;
-  padding: 10px;
-  border-radius: 10px;
-}
-
-.download-btn:active {
-  background-color: #00ff62;
-}
